Move 'handleSkill', 'addSkill', 'removeSkill' to separate files.
